[#91] DhcTopBar, DhcTopBarState 을 designsystem 모듈로 이동
